+++ build.gradle
@@ -47,6 +47,15 @@ shadowJar {
     archiveClassifier = null
 }
 
+jar {
+    manifest {
+        attributes(
+                'Main-Class': 'idkname.Main'
+        )
+    }
+}
+
+
 run{
     standardInput = System.in
 }
