+++ src/main/java/idkname/utility/TaskList.java
@@ -11,13 +11,28 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Iterator;
 
+/**
+ * Represents a list of tasks.
+ * Provides operations to add, delete, retrieve, and update tasks.
+ */
 public class TaskList implements Iterable<Task> {
     private ArrayList<Task> tasks = new ArrayList<>();
 
+    /**
+     * Constructs an empty TaskList.
+     */
     public TaskList() {
         this.tasks = new ArrayList<>();
     }
 
+    /**
+     * Adds a task to the list given its type and description.
+     * Supports task types: "todo", "deadline", and "event".
+     *
+     * @param type the type of task to add ("todo", "deadline", or "event")
+     * @param item the task description and additional data (e.g., date/time)
+     * @throws DateTimeException if the input format for a date/time is invalid
+     */
     public void add(String type, String item) {
         Task t = null;
         switch (type) {
@@ -55,10 +70,22 @@ public class TaskList implements Iterable<Task> {
 
     }
 
+    /**
+     * Adds a task directly to the list.
+     *
+     * @param t the task to add
+     */
     public void add(Task t) {
         this.tasks.add(t);
     }
 
+    /**
+     * Deletes a task from the list given its task number (1-based).
+     *
+     * @param taskNumber the 1-based task number as a string
+     * @throws NumberFormatException if the task number is not a valid integer
+     * @throws IndexOutOfBoundsException if the task number does not exist in the list
+     */
     public void delete(String taskNumber) {
         int task_Id = Parser.getTaskId(taskNumber); // may throw NumberFormatException
         Task t = this.tasks.get(task_Id);
@@ -69,24 +96,49 @@ public class TaskList implements Iterable<Task> {
                 this.tasks.size());
     }
 
+    /**
+     * Retrieves a task by its zero-based index.
+     *
+     * @param taskNumber the zero-based index of the task
+     * @return the task at the specified index
+     * @throws IndexOutOfBoundsException if the index is out of range
+     */
     public Task get(int taskNumber) {
         return this.tasks.get(taskNumber);
     }
 
+    /**
+     * Returns the underlying list of tasks.
+     *
+     * @return an ArrayList containing all tasks
+     */
     public ArrayList<Task> getTasks() {
         return this.tasks;
     }
 
+    /**
+     * Marks a task as done or undone, given its task number (1-based).
+     *
+     * @param b           true to mark as done, false to mark as undone
+     * @param taskNumber  the 1-based task number as a string
+     * @throws NumberFormatException if the task number is not a valid integer
+     * @throws IndexOutOfBoundsException if the task number does not exist in the list
+     */
     public void markDoneOrUndone(boolean b, String taskNumber) {
         int task_Id = Parser.getTaskId(taskNumber); // may throw NumberFormatException
         Task t = this.tasks.get(task_Id);
         if (b) {
-            t.markDone(true);
+            t.markDone(false);
         } else {
             t.markUndone();
         }
     }
 
+    /**
+     * Returns an iterator over the tasks in the list.
+     *
+     * @return an iterator of Task objects
+     */
     @Override
     public Iterator<Task> iterator() {
         return tasks.iterator();
