+++ src/main/java/idkname/utility/Ui.java
@@ -3,42 +3,76 @@ package idkname.utility;
 import java.time.DateTimeException;
 import java.util.Scanner;
 
+/**
+ * Handles all user interactions with the application, including
+ * greetings, goodbye messages, error messages, and command input.
+ */
 public class Ui {
     private final Scanner scanner;
     private final String name;
     private final String line = "_".repeat(75);
     private final TaskList tasks;
 
+    /**
+     * Constructs a Ui instance with a bot name and a reference to the task list.
+     *
+     * @param name  the name of the bot
+     * @param tasks the task list to operate on
+     */
     public Ui(String name, TaskList tasks) {
         this.scanner = new Scanner(System.in);
         this.name = name;
         this.tasks = tasks;
     }
 
+    /**
+     * Displays an error message if no save file is found when loading tasks.
+     */
     public void showFileLoadingError() {
         System.out.printf("No saved file.%n%s%n", this.line);
     }
 
+    /**
+     * Displays an error message if the user provides an invalid date or time format.
+     * The expected formats are:
+     * <ul>
+     *   <li>deadline description/yyyy-MM-dd</li>
+     *   <li>event description/yyyy-MM-ddTHH:mm:ss/yyyy-MM-ddTHH:mm:ss</li>
+     * </ul>
+     */
     public void showDateTimeError() {
         System.out.printf("Invalid date format.Please enter as:" +
                 "%n-deadline description/yyyy-MM-dd" +
                 "%n-event description/yyyy-MM-ddTHH:mm:ss/yyyy-MM-ddTHH:mm:ss%n%s%n", this.line);
     }
 
+    /**
+     * Displays an error message if task data cannot be saved to disk.
+     */
     public void showIOError() {
         System.out.printf("Unable to save data.%n%s%n", this.line);
     }
 
+    /**
+     * Displays an error message if the user specifies an invalid task index.
+     */
     public void showIndexOutOfBoundsError() {
         System.out.printf("Task not found. Please enter a valid task number.%n%s%n",
                 this.line);
     }
 
+    /**
+     * Displays an error message if the user provides an invalid number format
+     * (e.g. a non-integer where a task index is expected).
+     */
     public void showNumberFormatError() {
         System.out.printf("Please enter a valid command:" +
                 "%n[command] [task number]%n%s%n", this.line);
     }
 
+    /**
+     * Greets the user at the start of the program.
+     */
     public void greetings() {
         System.out.printf("%s" +
                 "%nHello! I'm %s" +
@@ -46,11 +80,30 @@ public class Ui {
                 "%n%s%n", this.line, this.name, this.line);
     }
 
+    /**
+     * Displays a goodbye message when the program exits.
+     */
     public void goodbye() {
         System.out.printf("Bye. Hope to see you again!" +
                 "%n%s", this.line);
     }
 
+    /**
+     * Main loop for handling user input.
+     * Reads commands from the user, parses them, and executes the corresponding actions.
+     * Supports commands such as:
+     * <ul>
+     *   <li><code>bye</code> — exit the program</li>
+     *   <li><code>list</code> — display all tasks</li>
+     *   <li><code>mark &lt;id&gt;</code> — mark a task as done</li>
+     *   <li><code>unmark &lt;id&gt;</code> — unmark a task</li>
+     *   <li><code>delete &lt;id&gt;</code> — delete a task</li>
+     *   <li><code>todo &lt;description&gt;</code> — add a todo task</li>
+     *   <li><code>deadline &lt;description&gt; /by yyyy-MM-dd</code> — add a deadline task</li>
+     *   <li><code>event &lt;description&gt; /from yyyy-MM-ddTHH:mm:ss /to yyyy-MM-ddTHH:mm:ss</code> — add an event task</li>
+     * </ul>
+     * Handles invalid commands and errors gracefully by showing appropriate error messages.
+     */
     public void echo() {
         String userInput;
         while (true) {
